+++ README.md
@@ -1,1 +1,4 @@
 # banking-app
+
+## Applied Patterns
+- Strategy Pattern: money_storage, vault
